requirements-dev.txt
gradio==3.41.2
hydra-core==1.3.2
langchain==0.0.273
openai==0.27.9
pinecone-client==2.2.2
python-dotenv==1.0.0
tiktoken==0.4.0